OVSDriver: move MAC lookup to ind_ovs_port_added
The previous location was historical, since the vport added multicast message
used to contain the MAC address.
